se creo PokemonRegion y se separo de App
